src/requirements.txt
pdfplumber==0.10.3
python-docx==1.1.0; python_version >= "3.14"
